Move execution_error to separate file
